requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch
transformers
bitsandbytes
python-multipart
pydantic
python-dotenv
